# JaegerExporter: Performance Improvements 2