… src/OpenTelemetry.Exporter.Jaeger/Implementation/JaegerUdpBatcher.cs
… src/OpenTelemetry/Trace/Internal/EvictingQueue.cs