More improvements for the Jaeger Exporter. Primarily, when enumerating over Lists/Dictionaries on structs, we will no longer pay a boxing penalty.

Before:
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT
  DefaultJob : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT


```
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev |      Median |       Gen 0 |     Gen 1 | Gen 2 |  Allocated | Completed Work Items | Lock Contentions |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|------------:|------------:|----------:|------:|-----------:|---------------------:|-----------------:|
| **JaegerExporter_Batching** |               **1** |         **10000** |    **151.6 ms** |   **3.80 ms** |  **11.20 ms** |    **148.5 ms** |   **1000.0000** |         **-** |     **-** |   **10.94 MB** |               **1.0000** |                **-** |
| **JaegerExporter_Batching** |              **10** |         **10000** |  **1,402.4 ms** |  **26.81 ms** |  **22.39 ms** |  **1,409.0 ms** |  **13000.0000** |         **-** |     **-** |   **107.3 MB** |               **2.0000** |                **-** |
| **JaegerExporter_Batching** |             **100** |         **10000** | **14,197.9 ms** | **355.47 ms** | **315.11 ms** | **14,134.7 ms** | **134000.0000** | **1000.0000** |     **-** | **1070.91 MB** |               **4.0000** |                **-** |

After:
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT
  DefaultJob : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT


```
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev |      Gen 0 |     Gen 1 | Gen 2 | Allocated | Completed Work Items | Lock Contentions |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|-----------:|----------:|------:|----------:|---------------------:|-----------------:|
| **JaegerExporter_Batching** |               **1** |         **10000** |    **146.5 ms** |   **2.92 ms** |   **4.63 ms** |          **-** |         **-** |     **-** |   **6.74 MB** |               **1.0000** |                **-** |
| **JaegerExporter_Batching** |              **10** |         **10000** |  **1,442.4 ms** |  **74.31 ms** |  **85.58 ms** |  **8000.0000** |         **-** |     **-** |   **65.3 MB** |               **2.0000** |                **-** |
| **JaegerExporter_Batching** |             **100** |         **10000** | **14,192.9 ms** | **216.25 ms** | **180.58 ms** | **81000.0000** | **1000.0000** |     **-** | **650.91 MB** |               **4.0000** |                **-** |

The majority of the remaining allocations are coming from the Thrift library when converting strings to UTF-8 byte arrays. I opened a [PR](https://github.com/apache/thrift/pull/2043) for that into the Apache repo. This PR combined with the Thrift change leads to the "zero-allocation"/flat growth target I was trying to achieve:
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT
  DefaultJob : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT


```
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 |  Allocated | Completed Work Items | Lock Contentions |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|------:|------:|------:|-----------:|---------------------:|-----------------:|
| **JaegerExporter_Batching** |               **1** |         **10000** |    **153.3 ms** |   **3.43 ms** |   **6.10 ms** |     **-** |     **-** |     **-** |  **256.76 KB** |               **1.0000** |                **-** |
| **JaegerExporter_Batching** |              **10** |         **10000** |  **1,511.4 ms** |  **29.43 ms** |  **41.26 ms** |     **-** |     **-** |     **-** |  **438.62 KB** |               **2.0000** |                **-** |
| **JaegerExporter_Batching** |             **100** |         **10000** | **15,289.7 ms** | **298.87 ms** | **332.20 ms** |     **-** |     **-** |     **-** | **2270.66 KB** |               **4.0000** |                **-** |

That's **KB** not MB!